% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_slider_color.R
\name{bootstrap_slider_color}
\alias{bootstrap_slider_color}
\title{Changes bootstrap slider input color individually.
Note there is currently no solution to change all of them simultaneously
Stick this function in a for loop for the range of slider_index arguments
to make this work}
\usage{
bootstrap_slider_color(slider_index = 0, colour = "mediumseagreen")
}
\arguments{
\item{slider_index}{A numeric of length one}

\item{colour}{A character of length one that takes names, rgb or rgba character vectors of rthe slider color}
}
\value{
A HTML head element incliuding the necessary css.
}
\description{
Changes bootstrap slider input color individually.
Note there is currently no solution to change all of them simultaneously
Stick this function in a for loop for the range of slider_index arguments
to make this work
}
\examples{
bootstrap_slider_color()
_
}
