% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_slider_colour.R
\name{bootstrap_slider_colour}
\alias{bootstrap_slider_colour}
\title{Changes bootstrap slider input color individually.
Note there is currently no solution to change all of them simultaneously
Stick this function in a for loop for the range of slider_index arguments
to make this work}
\usage{
bootstrap_slider_colour(slider_index = 0, colour = "mediumseagreen")
}
\arguments{
\item{slider_index}{A numeric of length one}

\item{colour}{A character of length one that takes names, rgb or rgba character vectors of rthe slider color}
}
\value{
A HTML head element incliuding the necessary css.
}
\description{
Changes bootstrap slider input color individually.
Note there is currently no solution to change all of them simultaneously
Stick this function in a for loop for the range of slider_index arguments
to make this work
}
<<<<<<< HEAD
=======
\examples{
bootstrap_slider_colour()
bootstrap_slider_colour(colour='blue')
}
>>>>>>> f1215f266c5877fc5419fb7f3880749e348db475
