% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_check.R
\name{chk_dups}
\alias{chk_dups}
\title{checks and diagnoses NA presence in vectors}
\usage{
chk_dups(x)
}
\arguments{
\item{x}{A vector of \emph{any} type}
}
\value{
A vector indicating duplicated entries, or length 0 vector if none
}
\description{
checks and diagnoses NA presence in vectors
}
\examples{
x <- c(3,4,5,6,6,7,7,8,3,9)
chk_dups(x)
}
