% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_check.R
\name{chk_nas}
\alias{chk_nas}
\title{checks and diagnoses NA presence in vectors}
\usage{
chk_nas(x)
}
\arguments{
\item{x}{A vector of \emph{any} type}
}
\value{
A numeric indicating NA entries, or length 0 vector if none
}
\description{
checks and diagnoses NA presence in vectors
}
\examples{
x <- c(3,4,5,6,NA,7,NA,8,NA,9)
chk_nas(x)
}
