% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_rows.R
\name{collapse_rows}
\alias{collapse_rows}
\title{collapse row values from a table into a list for pretty printing}
\usage{
collapse_rows(df, group_by_var, many_mapping_col)
}
\arguments{
\item{df}{dataframe, tibble, The data structure to work on}

\item{group_by_var}{tidy-eval, character The variable to group by}

\item{many_mapping_col}{The column of values that will collapse into a list.}
}
\value{
The inputs pasted together as a character string.
}
\description{
TThis function folds values of a many to one relationship together in a
single character composed of comma separated values.
}
\details{
The inputs can be anything that casn be input into
the paste function.
}
\note{
ensure the total values to fold are sufficiently low tp preserve the pretty print aesthetic
}
\section{experimant first if it helps}{

\subsection{not well supported }{
on multiple data types
}
}

\examples{

#example data 
nodes_tot=data.frame(
 FirstName=sample(c('A','B','C'),size = 50,replace=TRUE),
 LastName=sample(c('1','2','3'),size = 50,replace=TRUE),
 Latitude=sample(c('D','E','F'),size = 50,replace=TRUE),
 Longitude=sample(c('4','5','6'),size = 50,replace=TRUE))
#---------

collapse_rows(nodes_tot,FirstName,Longitude)
 
}
