% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\docType{data}
\name{custom_theme}
\alias{custom_theme}
\title{Themes a ggplot object}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 10.
}
\usage{
custom_theme
}
\value{
A themed \emph{ggplot} object
}
\description{
Themes a ggplot object
}
\examples{
if(!require(dplyr)){library(dplyr)}
if(!require(ggplot2)){library(ggplot2)}
mtcars <- mutate(mtcars,am=as.factor(am))
ggplot_obj <-
     ggplot( mtcars) + 
     facet_wrap(~am)+
     labs(title = 'Graph Title')+
     geom_point(aes(disp,mpg,col=mpg))
     
}
\keyword{datasets}
