% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoApp.R
\name{demoApp}
\alias{demoApp}
\title{Run a shiny appliaction demonstrating shiny functionality in RShinyHelpers}
\usage{
demoApp()
}
\description{
Run a shiny appliaction demonstrating shiny functionality in RShinyHelpers
}
\examples{
if (interactive()) {
demoApp()
}
}
