% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DT_selected_row_colour.R
\name{DT_peripheral_colour}
\alias{DT_peripheral_colour}
\title{Colors the colour of the peripheral element in a DT within a shiny app.  Call in the ui
element of a shiny app}
\usage{
DT_peripheral_colour(colour = "pink")
}
\arguments{
\item{colour}{A character vector of rgb,rgba,colour name}
}
\value{
A character vector of html style tag containing the relevant css
}
\description{
Colors the colour of the peripheral element in a DT within a shiny app.  Call in the ui
element of a shiny app
}
\examples{
DT_peripheral_colour()
DT_peripheral_colour(colour='blue')
}
