% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance_value.R
\name{finance_vbox}
\alias{finance_vbox}
\title{Changes bootstrap slider input color individually.
Note there is currently no solution to change all of them simultaneously
Stick this function in a for loop for the range of slider_index arguments
to make this work}
\usage{
finance_vbox(
  toplot = c(1, 2, 19, 20, 90, 100, 50, 200),
  label = "label",
  textid = "textOutput_here",
  col = "rgba(0,0,0,0.6)"
)
}
\arguments{
\item{toplot}{A numeric vector for the sparklines}

\item{label}{A character vector for length one for the box name}

\item{textid}{A character vector for the currency value.}

\item{col}{A character of length one that takes names, rgb or rgba character vectors of the box background}
}
\value{
A tagList of the div tree to be rendered to the shiny ui
}
\description{
Changes bootstrap slider input color individually.
Note there is currently no solution to change all of them simultaneously
Stick this function in a for loop for the range of slider_index arguments
to make this work
}
\examples{
finance_vbox(label = 'Important metric',textid = '$463')
}
