% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_format.R
\name{forCur}
\alias{forCur}
\title{Formats a numeric vector to the desired units and concommitant styling}
\usage{
forCur(x)
}
\arguments{
\item{x}{A numeric of any length}
}
\value{
A formated character string
}
\description{
Formats a numeric vector to the desired units and concommitant styling
}
\examples{
forCur(405556)
forCurGBP(346236)
forCurArea(452)
forCurAreaGBP(236)
}
