% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_value.R
\name{four_value}
\alias{four_value}
\title{builds the div for a more complicated multiple grouped text-value
boxes}
\usage{
four_value(text, value)
}
\arguments{
\item{text}{The labels -  character of length four}

\item{value}{The value - A character vector of length four}
}
\value{
A tagList of the div tree to be rendered in the ui of a  shiny app
}
\description{
builds the div for a more complicated multiple grouped text-value
boxes
}
\examples{

text=c('BTC','ETH','DOG','KLI')
values=c('$3,034','$4,353','$496','$629')
four_value(text,values)

}
