% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similar_permutations.R
\name{similar_permutations}
\alias{similar_permutations}
\title{find the distinct \emph{permutations} of columns}
\usage{
similar_permutations(edges, to, from)
}
\arguments{
\item{edges}{dataframe, tibble, The data structure to work on}

\item{to}{tidy-eval, character The variable to group by}

\item{from}{The column of values that will collapse into a list.}
}
\value{
The inputs pasted together as a character string.
}
\description{
find the distinct \emph{permutations} of columns, particularly in graph applications were
describing relationships between unordered nodes, or other applications were the order
does not matter
}
\details{
The inputs can be anything that casn be input into
the paste function.
}
\note{
ensure the total values to fold are sufficiently low tp preserve the pretty print aesthetic
}
\section{experimant first if it helps}{

\subsection{not well supported }{
on multiple data types
}
}

\examples{
if(!require(dplyr)){library(dplyr)}
edges <- data.frame(
  to=sample(c('A','B','C','D'),size = 50,replace=TRUE),
  from=sample(c('A','B','C','D','E','F'),size = 50,replace=TRUE))

dplyr::count(edges,to,from)

ordered = similar_permutations(edges,to,from)
 
dplyr::count(ordered, to ,from)
}
