% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbox.R
\name{vbox}
\alias{vbox}
\title{builds the div for a simple yet visually effective shiny value box}
\usage{
vbox(label = "label", textid = "textOutput_here", col = "rgba(60,130,180,0.5)")
}
\arguments{
\item{label}{The label -  character of length one}

\item{textid}{The value - A character vector of length one, optionally the textOutput of a reactive value}

\item{col}{Bkgrd col- A character of rbg(), rgba(), or a color name to}
}
\value{
A character vector of the div tree to be rendered in the ui of a  shiny app
}
\description{
builds the div for a simple yet visually effective shiny value box
}
\examples{
vbox()
}
