% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbox.R
\name{vbox_graphic}
\alias{vbox_graphic}
\title{builds the div for a shiny value box with a simple sparklines-like graphic}
\usage{
vbox_graphic(
  label = "label",
  textid = "textOutput_here",
  col = "rgba(60,130,180,0.5)",
  graph_title = "nowcast",
  value = 1:10
)
}
\arguments{
\item{label}{The label -  character of length one}

\item{textid}{The value - A character vector of length one, optionally the textOutput of a reactive value
if using vbox_reactive}

\item{col}{Bkgd col- A character of rbg(), rgba(), or a html color name  recognized in R}

\item{graph_title}{\itemize{
\item Sparkline title, character of length one
}}

\item{value}{the sparklines values to be plotted - a numeric vector}
}
\value{
A character vector of the div tree to be rendered in the ui of a shiny app
}
\description{
builds the div for a shiny value box with a simple sparklines-like graphic
}
\examples{
vbox_graphic()
}
