% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbox.R
\name{vbox_reactive}
\alias{vbox_reactive}
\title{builds the div for a simple yet visually effective shiny value box}
\usage{
vbox_reactive(
  label = "label",
  textid = "textOutput_here",
  col = "rgba(60,130,180,0.5)"
)
}
\arguments{
\item{label}{The value box header label to appear-  character of length one}

\item{textid}{The value - A character vector of length one, optionally the textOutput of a reactive value
if using vbox_reactive}

\item{col}{Bkgrd col- A character of rbg(), rgba(), or a chtml olor name to recognixed in R}
}
\value{
A character vector of the div tree to be rendered in the ui of a  shiny app
}
\description{
builds the div for a simple yet visually effective shiny value box
}
\examples{
vbox_reactive()

}
