% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{custom_theme}
\alias{custom_theme}
\title{Themes a ggplot object}
\usage{
custom_theme()
}
\value{
A themed \emph{ggplot} object
}
\description{
Themes a ggplot object
}
\examples{
# if(!require(dplyr)){library(dplyr)}
#if(!require(ggplot2)){library(ggplot2)}

mtcars <- dplyr::mutate(mtcars,am=as.factor(am))
ggplot_obj <-
     ggplot2::ggplot( mtcars) + 
     ggplot2::facet_wrap(~am)+
     ggplot2::labs(title = 'Graph Title')+
     ggplot2::geom_point(ggplot2::aes(disp,mpg,col=mpg))
     
ggplot_obj + custom_theme()

}
