% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_theme.R
\name{custom_theme}
\alias{custom_theme}
\title{Themes a ggplot object}
\usage{
custom_theme(ggplot_obj = NULL)
}
\arguments{
\item{ggplot}{A ggplot object}
}
\value{
A themed \emph{ggplot} object
}
\description{
Themes a ggplot object
}
\examples{
ggplot_obj <- mtcars\%>\%
              mutate(am=as.factor(am))\%>\%
              ggplot( ) + 
     facet_wrap(~am)+
     labs(title = 'Graph Title')+
     geom_point(aes(disp,mpg,col=mpg))
     
}
